updated _zm_ai_dogs.gsc fixing a rare crash more in description
Continues in for loops seem to have a weird tendency to overflow causing an infinite loop. This was fixed by changing it to a while loop.  Additionally logprints() have been added to make up for the lack of developer mode. Use the _zm_bot.gsc to at the same time as _zm_ai_dogs.gsc to use it.