Update _zm_challenges.gsc
did a quick check of this to see if i could get it to compile. probably screwed up somewhere, because my compiler crashes with a System.ArgumentOutOfRangeException. (using the compiler.exe from the plutonium forums), and gsc studio downloaded from a link in this repo just gets stuck exporting.  issue is most likely in the areas that don't completely match due to it causing a compiler error if matched, if i had to guess.